src/backend/requirements.txt
Django==4.2
django-cors-headers==4.2
dj-database-url==2.0
django-cache-url==3.4
django-cacheops==7.0
whitenoise==6.5
django-taggit==4.0
pyjwt==2.8
psycopg2-binary==2.9.6
Pillow==10.0
django-filter==23.2
git+https://github.com/TcMits/djview.git@v0.2.0
apispec==6.3
bjoern==3.2.2
